[PYTHON] Added setup.py options to run benchmark suite
