+++ python/test/test_cross_entropy.py
