+++ python/triton/__init__.py
@@ -2,6 +2,7 @@
 # or pybind11 shows `munmap_chunk(): invalid pointer`
 import torch
 # submodules
+from . import testing
 from .kernel import *
 from . import ops
 # C bindings
